Add SubColumn float encoding support (#851)
This syncs the SubColumn implementation used by the paper artifact, including FLOAT scaling support and round-trip tests.

+++ cpp/src/encoding/gorilla_encoder.h
@@ -111,7 +111,7 @@ class GorillaEncoder : public Encoder {
                 bits_left_ = 0;
             } else {
                 shift = bits_left_ - bits;
-                buffer_ |= (uint8_t)(value << shift);
+                buffer_ |= (uint8_t)(static_cast<uint64_t>(value) << shift);
                 bits_left_ -= bits;
                 bits = 0;
             }
